Break up sweet_conductor into 3 files
