+++ crates/holochain/src/test_utils/sweettest/mod.rs
@@ -16,6 +16,8 @@ mod sweet_agents;
 mod sweet_app;
 mod sweet_cell;
 mod sweet_conductor;
+mod sweet_conductor_batch;
+mod sweet_conductor_handle;
 mod sweet_dna;
 mod sweet_network;
 mod sweet_zome;
@@ -24,6 +26,8 @@ pub use sweet_agents::*;
 pub use sweet_app::*;
 pub use sweet_cell::*;
 pub use sweet_conductor::*;
+pub use sweet_conductor_batch::*;
+pub use sweet_conductor_handle::*;
 pub use sweet_dna::*;
 pub use sweet_network::*;
 pub use sweet_zome::*;
