+++ crates/holochain/src/test_utils/sweettest/sweet_conductor_batch.rs
@@ -1,0 +1,120 @@
+use super::{standard_config, SweetAgents, SweetAppBatch, SweetConductor};
+use crate::conductor::config::ConductorConfig;
+use futures::future;
+use hdk::prelude::*;
+use holochain_types::prelude::*;
+
+/// A collection of SweetConductors, with methods for operating on the entire collection
+#[derive(derive_more::From, derive_more::Into, derive_more::IntoIterator)]
+pub struct SweetConductorBatch(Vec<SweetConductor>);
+
+impl SweetConductorBatch {
+    /// Map the given ConductorConfigs into SweetConductors, each with its own new TestEnvironments
+    pub async fn from_configs<I: IntoIterator<Item = ConductorConfig>>(
+        configs: I,
+    ) -> SweetConductorBatch {
+        future::join_all(configs.into_iter().map(SweetConductor::from_config))
+            .await
+            .into()
+    }
+
+    /// Create the given number of new SweetConductors, each with its own new TestEnvironments
+    pub async fn from_config(num: usize, config: ConductorConfig) -> SweetConductorBatch {
+        Self::from_configs(std::iter::repeat(config).take(num)).await
+    }
+
+    /// Create the given number of new SweetConductors, each with its own new TestEnvironments
+    pub async fn from_standard_config(num: usize) -> SweetConductorBatch {
+        Self::from_configs(std::iter::repeat_with(standard_config).take(num)).await
+    }
+
+    /// Get the underlying data
+    pub fn iter(&self) -> impl Iterator<Item = &SweetConductor> {
+        self.0.iter()
+    }
+
+    /// Get the underlying data
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut SweetConductor> {
+        self.0.iter_mut()
+    }
+
+    /// Get the underlying data
+    pub fn into_inner(self) -> Vec<SweetConductor> {
+        self.0
+    }
+
+    /// Opinionated app setup.
+    /// Creates one app on each Conductor in this batch, creating a new AgentPubKey for each.
+    /// The created AgentPubKeys can be retrieved via each SweetApp.
+    pub async fn setup_app(
+        &mut self,
+        installed_app_id: &str,
+        dna_files: &[DnaFile],
+    ) -> SweetAppBatch {
+        let apps = self
+            .0
+            .iter_mut()
+            .map(|conductor| async move {
+                let agent = SweetAgents::one(conductor.keystore()).await;
+                conductor
+                    .setup_app_for_agent(installed_app_id, agent, dna_files)
+                    .await
+            })
+            .collect::<Vec<_>>();
+
+        future::join_all(apps).await.into()
+    }
+
+    /// Opinionated app setup. Creates one app on each Conductor in this batch,
+    /// using the given agents and DnaFiles.
+    ///
+    /// The number of Agents passed in must be the same as the number of Conductors
+    /// in this batch. Each Agent will be used to create one app on one Conductor,
+    /// hence the "zipped" in the function name
+    ///
+    /// Returns a batch of SweetApps, sorted in the same order as the Conductors in
+    /// this batch.
+    pub async fn setup_app_for_zipped_agents(
+        &mut self,
+        installed_app_id: &str,
+        agents: &[AgentPubKey],
+        dna_files: &[DnaFile],
+    ) -> SweetAppBatch {
+        if agents.len() != self.0.len() {
+            panic!(
+                "setup_app_for_zipped_agents must take as many Agents as there are Conductors in this batch."
+            )
+        }
+
+        let apps = self
+            .0
+            .iter_mut()
+            .zip(agents.iter())
+            .map(|(conductor, agent)| {
+                conductor.setup_app_for_agent(installed_app_id, agent.clone(), dna_files)
+            })
+            .collect::<Vec<_>>();
+
+        future::join_all(apps).await.into()
+    }
+
+    /// Let each conductor know about each others' agents so they can do networking
+    pub async fn exchange_peer_info(&self) {
+        let envs = self.0.iter().map(|c| c.envs().p2p()).collect();
+        crate::conductor::p2p_store::exchange_peer_info(envs);
+    }
+}
+
+impl std::ops::Index<usize> for SweetConductorBatch {
+    type Output = SweetConductor;
+
+    fn index(&self, index: usize) -> &Self::Output {
+        &self.0[index]
+    }
+}
+
+impl std::ops::IndexMut<usize> for SweetConductorBatch {
+    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
+        &mut self.0[index]
+    }
+}
